Improve notes-ui (#1291)
* improve notes-ui

* don't attempt non-sponsor PUT

* lint

* fix logic

+++ js/hole-tabs.tsx
@@ -16,6 +16,7 @@ import {
     updateLocalStorage,
     getLang,
     setState,
+    ctrlEnter,
 } from './_hole-common';
 import { highlightCodeBlocks } from './_wiki';
 
@@ -219,7 +220,14 @@ layout.registerComponentFactoryFunction('code', async container => {
     };
 
     // Wire submit to clicking a button and a keyboard shortcut.
-    $('#runBtn').onclick = () => submit(editor, updateReadonlyPanels);
+    const closuredSubmit = () => submit(editor, updateReadonlyPanels);
+    $('#runBtn').onclick = closuredSubmit;
+    $('#editor').onkeydown = ctrlEnter(closuredSubmit);
+    $('#holeLangNotesEditor-section').onkeydown = ctrlEnter(async () => {
+        if (await convertNotesAndRun()) {
+            await upsertNotes();
+        }
+    });
 
     const deleteBtn = $('#deleteBtn');
     if (deleteBtn) {
@@ -233,12 +241,14 @@ layout.registerComponentFactoryFunction('code', async container => {
 async function upsertNotes() {
     $<HTMLButtonElement>('#upsert-notes-btn').disabled = true;
     const content = holeLangNotesEditor!.state.doc.toString();
-    const resp = await fetch(
-        `/api/notes/${hole}/${getLang()}`,
-        content ? { method: 'PUT', body: content} : { method: 'DELETE' },
-    );
-    if (resp.status !== 204) $<HTMLButtonElement>('#upsert-notes-btn').disabled = false;
-    else holeLangNotesContent = content;
+    if (!content || isSponsor()) {
+        const resp = await fetch(
+            `/api/notes/${hole}/${getLang()}`,
+            content ? { method: 'PUT', body: content} : { method: 'DELETE' },
+        );
+        if (resp.status !== 204) $<HTMLButtonElement>('#upsert-notes-btn').disabled = false;
+        else holeLangNotesContent = content;
+    }
 };
 
 function parseSubstitutions() {
@@ -252,15 +262,19 @@ function parseSubstitutions() {
     $<HTMLButtonElement>('#convert-notes-btn').disabled = substitutions.length < 1;
 }
 
-function convertNotesAndRun(){
+async function convertNotesAndRun(): Promise<boolean> {
     if (editor && holeLangNotesEditor) {
-        let notes = holeLangNotesEditor.state.doc.toString();
+        let newCode = holeLangNotesEditor.state.doc.toString();
         for (const {pattern, replacement} of substitutions) {
-            notes = notes[pattern.global ? 'replaceAll' : 'replace'](pattern, replacement);
+            newCode = newCode[pattern.global ? 'replaceAll' : 'replace'](pattern, replacement);
         }
-        setState(notes, editor);
-        submit(editor, updateReadonlyPanels);
+        const code = editor.state.doc.toString();
+        if (code !== newCode) {
+            setState(newCode, editor);
+            return await submit(editor, updateReadonlyPanels);
+        }
     }
+    return false;
 }
 
 layout.registerComponentFactoryFunction('holeLangNotes', async container => {
@@ -286,6 +300,7 @@ layout.registerComponentFactoryFunction('holeLangNotes', async container => {
     $('#convert-notes-btn').onclick = convertNotesAndRun;
     $('#notes-substitutions').oninput = parseSubstitutions;
 
+    await afterDOM();
     updateHoleLangNotesContent();
     parseSubstitutions();
 });
@@ -401,7 +416,7 @@ function isSponsor() {
     return $('#golferInfo')?.dataset.isSponsor === 'true';
 }
 function hasNotes() {
-    return $('#golferInfo')?.dataset.hasNotes  === 'true';
+    return $('#golferInfo')?.dataset.hasNotes === 'true';
 }
 
 function getPoolFromLayoutConfig(config: LayoutConfig) {
